# next-sample/backend/application/api/openapi.yaml
openapi: "3.0.3"
info:
  version: 0.0.1
  title: REST API
servers:
  - url: http://localhost:9000
tags:
  - name: recipe
  - name: version
paths:
  /:
    get:
      summary: "バージョン確認"
      operationId: getVersion
      tags:
        - version
      responses:
        "200":
          description: OK
  /recipe/{recipeId}:
    get:
      summary: "レシピ取得"
      operationId: getRecipe
      tags:
        - recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          description: レシピIDを指定
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Recipe"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                unexpected_error:
                  $ref: "#/components/examples/unexpected_error"
  /recipe/list:
    get:
      summary: "レシピ一覧取得"
      operationId: getRecipeList
      tags:
        - recipe
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeList"
  /recipe/register:
    post:
      summary: "新規レシピ登録"
      operationId: registerRecipe
      tags:
        - recipe
      responses:
        "201":
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              items:
                  $ref: "#/components/schemas/Recipe"
components:
  schemas:
    Recipe:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: レシピID
          example: レシピID
        title:
          type: string
          description: レシピタイトルを20文字で指定できます
          example: レシピ名
        materials:
          type: array
          description: 最大10個の材料を15文字で指定できます
          items:
            type: string
          example:
            - 白菜
            - トマト
            - きゅうり
        content:
          type: string
          description: 作り方を500文字まで設定できます
          example: 1.白菜を食べやすい大きさに切る\n2.白菜を茹でる
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - materials
        - content
        - created_at
        - updated_at
    RecipeList:
      type: object
      properties:
        total_count:
          type: integer
        per_page:
          type: integer
        page:
          type: integer
        list:
          type: array
          items:
            $ref: "#/components/schemas/Recipe"
      required:
        - total_count
        - per_page
        - page
        - list
    error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  examples:
    unexpected_error:
      summary: Unexpected error
      value: |-
        {
          "code": "ERR90002",
          "message": "We couldn't handle your request."
        }