# next-sample/backend/application/infrastructure/openapi/model.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package openapi

import (
	"time"
)

// Recipe defines model for Recipe.
type Recipe struct {
	// Content 作り方を500文字まで設定できます
	Content   string    `json:"content"`
	CreatedAt time.Time `json:"created_at"`

	// Id レシピID
	Id int64 `json:"id"`

	// Materials 最大10個の材料を15文字で指定できます
	Materials []string `json:"materials"`

	// Title レシピタイトルを20文字で指定できます
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updated_at"`
}

// RecipeList defines model for RecipeList.
type RecipeList struct {
	List       []Recipe `json:"list"`
	Page       int      `json:"page"`
	PerPage    int      `json:"per_page"`
	TotalCount int      `json:"total_count"`
}

// Error defines model for error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// GetRecipeListParams defines parameters for GetRecipeList.
type GetRecipeListParams struct {
	Page *int `form:"page,omitempty" json:"page,omitempty"`
}

// RegisterRecipeJSONBody defines parameters for RegisterRecipe.
type RegisterRecipeJSONBody = map[string]interface{}

// RegisterRecipeJSONRequestBody defines body for RegisterRecipe for application/json ContentType.
type RegisterRecipeJSONRequestBody = RegisterRecipeJSONBody